requirements.txt
absl-py==2.1.0
actionlib==1.14.0
angles==1.9.13
apturl==0.5.2
astunparse==1.6.3
attrs==24.3.0
autobahn==23.1.2
Automat==24.8.1
base_local_planner==1.17.3
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.5.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==23.10.4
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cramjam==2.9.1
cryptography==44.0.0
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.11
dbus-python==1.2.16
deepface==0.0.93
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
distro-info==1.0
docutils==0.20.1
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.3
empy==4.2
entrypoints==0.4
fasteners==0.19
fastimport==0.9.8
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.3
future==1.0.0
gast==0.4.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gdown==5.2.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.13.1
grpcio==1.69.0
gunicorn==23.0.0
h5py==3.11.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
image-geometry==1.16.2
importlib_metadata==8.5.0
incremental==24.7.2
interactive-markers==1.12.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.5
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.13.1
keyring==25.5.0
kiwisolver==1.4.7
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.7
libclang==18.1.1
lockfile==0.12.2
louis==3.12.0
lz4==4.3.3
macaroonbakery==1.3.4
Mako==1.3.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.17.0
monotonic==1.6
mpi4py==4.0.1
mtcnn==0.1.1
netifaces==0.11.0
nose==1.3.7
numpy==1.24.3
oauthlib==3.2.2
olefile==0.47
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
paramiko==2.6.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pocketsphinx==5.0.4
protobuf==4.25.5
psutil==6.1.1
ptyprocess==0.7.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.26.1
pycparser==2.22
pycrypto==2.6.1
pycryptodomex==3.21.0
pycups==1.9.73
pydot==3.0.4
PyGithub==1.43.7
Pygments==2.19.1
PyGObject==3.36.0
PyHamcrest==2.1.0
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.7
pyOpenSSL==25.0.0
pyparsing==3.1.4
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.15.0
pyRFC3339==2.0.1
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-gitlab==2.0.1
python-gnupg==0.5.4
python-qt-binding==0.4.4
python-snappy==0.7.3
pytools==2024.1.14
PyTrie==0.4.0
pyttsx3==2.98
pytz==2024.2
pyxdg==0.28
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-unixsocket==0.3.0
resource_retriever==1.12.7
retina-face==0.0.17
roman==4.2
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.0
rosserial_python==0.9.2
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.25
SecretStorage==3.3.3
sensor-msgs==1.13.1
service-identity==24.2.0
simplejson==3.19.3
sip==6.8.6
six==1.17.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sortedcontainers==2.4.0
sound-play==0.3.18
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.10.4
srt==3.5.3
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tomli==2.2.1
topic-tools==1.17.0
tqdm==4.67.1
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2024.2
u-msgpack-python==2.8.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
vcstools==0.1.42
vosk==0.3.45
wadllib==2.0.0
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.1
wsaccel==0.6.6
wstool==0.1.18
xacro==1.14.19
xkit==0.0.0
zipp==3.20.2
zope.interface==7.2
